Tag browser now searches in a different thread and reuses previous search results when possible.
