+++ superhotmcc/src/haloce/TagBrowser.hpp
@@ -1,0 +1,4 @@
+namespace HaloCE::Mod::UI {
+    extern bool showTagBrowser;
+    void tagBrowser();
+}
